+++ CHANGES.rst
@@ -24,6 +24,7 @@ Unreleased
 - Remove ``Reddit.validate_on_submit`` configuration attribute.
 - Remove ``WebSocketException.original_exception`` method.
 - Remove the ``after`` argument for :meth:`.conversations`.
+- Remove ability to use :class:`.CommentForest` as an asynchronous iterator.
 
 7.8.1 (2024/12/21)
 ------------------